refactor: get rid of circular js imports
