+++ packages/apps/reputation-oracle/server/src/modules/qualification/user-qualification.entity.ts
@@ -1,21 +1,17 @@
 import { Entity, ManyToOne } from 'typeorm';
 import { DATABASE_SCHEMA_NAME } from '../../common/constants';
 import { BaseEntity } from '../../database/base.entity';
-/**
- * TODO: Leave fix follow-up refactoring
- * Importing from '../user' causes circular import error here.
- */
-import { UserEntity } from '../user/user.entity';
-import { QualificationEntity } from '../qualification/qualification.entity';
+import type { UserEntity } from '../user';
+import type { QualificationEntity } from '../qualification/qualification.entity';
 
 @Entity({ schema: DATABASE_SCHEMA_NAME, name: 'user_qualifications' })
 export class UserQualificationEntity extends BaseEntity {
-  @ManyToOne(() => UserEntity, (user) => user.userQualifications)
+  @ManyToOne('UserEntity', (user: UserEntity) => user.userQualifications)
   public user: UserEntity;
 
   @ManyToOne(
-    () => QualificationEntity,
-    (qualification) => qualification.userQualifications,
+    'QualificationEntity',
+    (qualification: QualificationEntity) => qualification.userQualifications,
   )
   public qualification: QualificationEntity;
 }
